Вынесли структуру Vector2D отдельно в пространство имен.
